model reconfig

+++ IslandFoodmart/Models/Payment.cs
@@ -5,7 +5,7 @@ namespace IslandFoodmart.Models
     public class Payment
     {
         public int PaymentID { get; set; }
-        public int PaymentAmount { get; set; }
+        public decimal PaymentAmount { get; set; }
         public string PaymentMethod { get; set; }
         public DateTime PaymentDate { get; set; }
         public ICollection<DatabaseUser> DatabaseUser { get; set; }
